Bug#34503695: Hypergraph: mysqld crash-signal 11               -CommonSubexpressionElimination
For a query which is modified by in2exists transformation,
hypergraph plans twice. One with materialization
and one without. So, before the hypergraph optimizer is
invoked for the case without materialization, a where
condition without predicates added by the in2exists
transformation is made from the actual where condition.
This would be used for the second time planning (with
materialzation).
For the case in bugpage, first planning would result in
an always false condition. This results in modification
of the original where condition which also results
in the modification of the where condition used for second
planning leading to problems.
E.g.
For a where condition like this one:
((t4.f1 <> t3.f1) or (t4.f1 < t3.f1)) and
<if>(outer_field_is_not_null, ((<cache>(t1.f1)
= t2.f1) or (t2.f1 is null)), true) and (t4.f1 = t3.f1))

In EarlyNormalizeConditions, (t4.f1 <> t3.f1) or (t4.f1 < t3.f1)
is changed to (t4.f1 <> t4.f1) or (t4.f1 < t4.f1) because
t4.f1 = t3.f1. After this, call to remove_eq_conds()
concludes that the condition would be always false. While
at it, it removes elements from the list for the OR condition.
Later, all elements from the entire array of conditions
created by hypergraph optimizer are removed and the list is
replaced with an always false condition. However, the original
where condition is transformed into this:
(() and <if>(outer_field_is_not_null, ((<cache>(t1.f1) = t2.f1)
 or (t2.f1 is null)), true) and (t4.f1 = t3.f1))
Note the missing elements in the first argument to the AND
condition. This is not a problem for most queries, because the
original where condition is not used in the rest of the query
processing. However, because the query block was modified by
in2exists transformation, hypergraph reuses the condition to
replan and this time, it finds an empty OR and crashes.

Solution is to make a copy of the and/or structure of the
where condition when in2exists transformation is present.
This will ensure that the removal of list elements in the
first planning does not affect the second one.

Change-Id: I9ab4c3ef4c191099632b1849567da16fcceea9cb