WL#15008: Remove Abstract Query Plan (AQP) interface
Previous patch had to mask out all tables in nests prior
to the 'root' as we prepared for reusing the same set of m_tables[]
for multiple root.

This patch removes that cludge, and instead fix the single
issue where we made the incorrect assumption that there would
not be any tables set in the nest-maps before the root table.

Change-Id: Ie3d1269dbe41a4bcb393564104bf568e547b4b12