Bug#34504569 Processing of ThreadConfig changed in 8.0.23 release
Remove the need to specify the mandatory threads.
main, ldm, recv and rep are now implicit.
If not specified 1 instance of each will be created
implicitly.
To have zero main, ldm or rep one needs to explicitly
set count=0 in the threadConfig string.
recv min count is 1.

If user forces count=0 an error is fired since it is
not allowed.
(same behaviour as in 8.0.30)

Change-Id: I6a1e2e57712c84c6d5e9d5ad1c43b8ae111ff257