WL#15008: Remove Abstract Query Plan (AQP) interface
Introduce ../storage/ndb/plugin/ha_query_plan.cc

ha_query_plan.cc is intended to contain the query-plan extraction
from the AccessPath structure, the setup of the member variables in
the struct pushed_table describing the 'plan', as well as the get'ters
for the 'plan'. It will eventually obsolete abstract_query_plan.*

In this patch we move over the get'ters working on the m_path, m_table,
and m_filter members introduced in the previous patch, as well as the
m_properties member. Thus getting a 'real' implementation of these methods
instead of the temporarykludge-implementation of calling m_aqp-><method_name>().

The existing implementation of these get'ters in the Table_access class
becomes obsolete and is removed. Also remove Table_access::get_item_equal()
which was already unused.

It will provide the same functionality as the existing AQP interface,
just using the ha_ndbcluster_push.h structures directly.

Change-Id: I271a50ee43f0e952cd2eb3d366bfcf26cbd37cf6