Bug#34427559 Quarantine mechanism affects establishing connections to Router
The current implementation uses the io_context timer as a trigger for
quarantined destination check. When it expires it calls a function that
checks the destination availability in context of this timer. This is a
blocking function that can block until the connect timeout expired.
This whole time the whole io_context is blocked.

This patch changes the check mechanism to use asynchronous IO the same
way the client connection requests are handled by the Router.

Change-Id: I40ca0dcb2de973847126842c85fd22906602e852