BUG#34066605: LOST+Found directory exists after upgrade to               8.0
Description:
============
Manually created directories in a 5.7 data directory with
special characters in the directory name (for eg:
lost+found) or schemas created with special characters with
`#mysql50#` prefix cause problems if the server is upgraded
to 8.0.

Analysis:
=========
5.7 allows users to manually create directories in the data
directory, which will be treated as schemas. Users are also
allowed to create schemas/tables with old-encoding i.e names
containing special characters, by adding a prefix
`#mysql50#` to the name. These are disallowed in 8.0.

However, if we create such schemas in 5.7 and upgrade to
8.0, the upgrade process creates entry in mysql.schemata for
all the folders found in data directory. While parsing the
directory name, if it finds a invalid character, it
hard-codes the character(s) with '?', logs an error to the
error log mentioning about the invalid name and stores the
new name in DD. The upgrade goes through successfully. Since
the directory name in the data directory and the name stored
in DD are not the same, any DDL on that schema will report
an error.

Fix:
====
During the upgrade, if it encounters a invalid directory
name/filename for a schema/table, the upgrade is stopped.

Note:
=====
In 5.7, to update old names to use the new encoding,
re-encode them by using mysqlcheck. The following commands
update all names to the new encoding:

mysqlcheck --check-upgrade --all-databases
mysqlcheck --fix-db-names --fix-table-names --all-databases

OR via mysql client, for eg:
ALTER DATABASE `#mysql50#lost+found` UPGRADE DATA DIRECTORY
NAME;

Change-Id: I630fcc72fca555d684a59b9ffe86685e6e7e7526