Bug#34534373: Heatwave offload issue - Sees inner tables of a               semijoin when it should not
While canonicalizing conditions, post pushdown, fields in
conditions were getting replaced with fields from inner
tables of a semijoin if there are multiple equalities.
This is not correct. So, we make sure that we use only
visible tables for the relational expression rather
than all tables used in that expression for replacements.
Also, during equality propagation in EarlyNormalizeConditions,
for a degenerate join condition i.e if the fields in the
join condition come from only one side of the join, we find
replacement from the same side of the join so that it can
be pushed down later.

Change-Id: I81869660dbbb583933e92f8b2b683160ddc98a92