Bug#34538071 NDBXFRM1 reader wrongly trimmed trailing non-zeros for header and trailer
Fixup 8.0.22: WL#13499 ndb_az31, ndb_ndbxfrm1: File format classes

The intention has always been that NDBXFRM1 reader will trim the
trailing zeros from header and trailer, but the bug caused the opposite
trimming non zeros.

This has not been a problem in 8.0.30 or earlier since the last field
in header was a zero word padding, and for trailer the last field was
for data checksum that was never checked and the next last field was
file checksum that was never written and always zero.

Change-Id: If6bdea3fab6d070aeb563697891dd902b38bdd03