Bug #34094706 : require_secure_transport causes abrupt server exit when checking                 thd->active_vio
Description:
------------
When require_secure_transport is enabled, the server checks the thd->active_vio
pointer in acl_authenticate(). However, this check is not thread-safe as server
can exit abruptly if the connection is killed before checking active_vio.

Fix:
----
Provided mutex protection in the acl_authenticate() method to avoid this problem
and added a nullptr check as well for thd->active_vio.

Change-Id: I42c284dcaf3c988734010f9faf02f9c023dda90d