Bug#34540660: Make straight joins as not associative for secondary engine
Make straight joins as not assciative for secondary engine. It expects that
the plans generated by hypergraph optimizer with STRAIGHT_JOIN hints to be
the same as that of old optimizer. Since old optimizer was only capable of
generating left-deep plans, heatwave expects similar plans. Since its seeing
regression for TPC-DS query - 24, for now we are making STRAIGHT_JOINs as
not associative for secondary engine.

Change-Id: I3eb5cb0025de951c3e6d3f2ebf1eda229970fb41