WL#15008: Remove Abstract Query Plan (AQP) interface
Let ndb_pushed_builder_ctx::construct() collect the 'm_inner_map'
of all tables being members of the same (inner*) Join_nest.
This can easily be collected directly into the Join_nest structure,
which is an integrated part of the ndb_pushed_builder_ctx, instead
of builing it afterwards from the first_ / last_inner.

Introduce pushed_table get'ters methods for returning these
'inner' tables:

  - get_full_inner_nest()
  - get_inner_nest(uint last)

Which returns a 'map' of respectively all table in inner nest
which this table is member of, and all tables prior to the specified 'last'

This replace the former function full_inner_nest() which has to calculate
this table_map by iterating the m_tables[] over the first_/last_inner
each time it was called.

Change-Id: I589dd007d6754ee9fa6e91bf26f8c8a258c50146