Bug #34501054 Evaluate the utility of SN to LSN conversion
Per Kuba's investigation and in his own words:
`free_check_limit_sn` is unnecessarily expressed in terms
of SN instead of LSN, which not only complicates the whole
code around it by needless back-and-forth conversions, but
also ties checkpointing to InnoDB redo log implementation
detail (as SN is InnoDB-specific). I've check with
https://godbolt.org/z/9T8x7brnG that compiler is smart enough
to eliminate divisions/modulo in either approach, so the performance aspect
isn't important IMO. But the code actually is shorter, and easier
to understand. Following is the resulting assembly code for quick reference:

log_translate_sn_to_lsn(unsigned long long):
  movabsq $595056260442243601, %rax
  movq %rdi, %rdx
  shrq $4, %rdx
  mulq %rdx
  movq %rdx, %rcx
  movq %rdx, %rax
  salq $5, %rcx
  salq $9, %rax
  subq %rdx, %rcx
  salq $4, %rcx
  subq %rcx, %rdi
  leaq 12(%rax,%rdi), %rax
  ret

Change-Id: I89683401bd7c02bbc65a9b200071918e5c5588ed