Bug#33725415: A bug in sql_lex.h cause mysql-server failure
This patch implements a new way to avoid removing item trees that
have multiple references. The old approach has been to identify
possible situations where an item tree to be removed is also
referenced from another location, thus removing the tree will cause
pointers to point at non-existing objects.

The new approach is to implement a reference count in every Item
object. It is incremented by one for every Item_ref object that points
to this Item. It is also incremented by one for every expression in
a select list, since only such items can also be referenced from
Item_ref objects.

Whenever an item tree is attempted to be removed, when descending down
the item tree and passing an Item_ref object, its referenced Item's
reference count is decremented by one, and only when the reference count
reaches zero is the item tree below the Item_ref deleted.

- The first change is to make Item::ref protected, rename the member to
  m_ref_item and implement a public accessor function ref_item().
  A helper function ref_pointer() that returns the value of m_ref_item
  is also added. This encapsulates the item reference pointer and makes
  it simpler to implement reference counting later.

- Two new Item functions increment_ref_count() and decrement_ref_count()
  are implemented.

- In Item_ref constructors, when another Item object is referenced, its
  reference count is incremented by one. Incrementing must also be
  performed in a couple of resolver functions.

- A new function Item_ref::clean_up_after_removal() is implemented.
  Its context object, Cleanup_after_removal_context, now inherits from
  class Item_tree_walker, so that it is possible to traverse a tree
  without deleting parts of it. Another field Item_ref::m_unlinked
  is also added so that we can distinguish between calling ::walk()
  going down or up.

- Another MARKER constant is implemented so that we know when reaching
  an item to be deleted going down or up. It helps to ensure that
  actual deletion happens postfix. This is only needed for
  Item_subselect objects, which contain query expressions objects
  that must be removed.

With reference counting in place, we can remove old code that was
used to avoid removing query expressions that were redundant, but had
a second reference that made it impossible:

- Item_cond::fix_fields() used to not remove item trees that contained
  Item_view_ref objects. Ref-counting will now handle this.

- Item_subselect::clean_up_after_removal() used to check that a subquery
  in a select list was resolved using an alias. Ref-counting will handle
  this case too.

- Item_subselect::clean_up_after_removal() also used to check that a
  subquery is not referenced from higher up in the item tree.

- In Query_block::convert_subquery_to_semijoin(), there is no longer
  need for special code for select list expressions in query
  expressions of EXISTS subqueries.

- Query_block::empty_order_list() used to not remove item trees that
  contained Item_view_ref objects. Ref-counting will now handle this.

Change-Id: Ifc89d179fd38a9b25607f8499423a2e11222bdf4