Bug #34513359 Hypergraph: Cost for field=dependent_subquery not included
This patch ensures that we include the cost of the subquery in the overall
cost in queries like:

SELECT <projection> FROM <tables>
WHERE field <comparison operator> <subquery>

'comparison operator' is '=','<' etc. and 'subquery' may be dependent or
independent of 'tables'. Prior to this fix, only the cost of subqueries
with IN/ALL/ANY were included (as well as the cost of subqueries
transformed into joins).

We now add the cost of Item_singlerow_subselect subqueries for each
filtered row for dependent (or otherwise uncachable) subqueries, and
to the filter init_cost for independent subqueries.

In addition, this patch fixes a releated case. Previously,
the cost of the subquery would not be added to the total for queries like:

SELECT <aggregate> FROM <tables> GROUP BY <group term>
HAVING <aggregated term> <comparison operator> <subquery>

This is now handled correctly.

Change-Id: Ia4d42f770cd0e75d8414635aec5d16e074b51b45