Bug#34085026: Hypergraph error mismatch where query ran in MySQL
HeatWave currently sets the row estimate to zero in all access paths.
This makes the max_join_size variable ineffective when executing
queries in HeatWave, since every query plan gives the impression of
accessing zero rows.

This patch works around the missing information by using the handler's
row estimate for scans of base tables.

Since the workaround only provides row estimates for the leaves in the
query plan, it is not possible to track the effects of LIMIT through
joins and filters, as we don't have any estimates for the cardinality
before and after the joins and filters and cannot calculate the
fanout. Queries with LIMIT may therefore raise errors when executing
in HeatWave and a max_join_size is in effect, even though the limit
would have prevented the large table scans from accessing more than
max_join_size rows.

Change-Id: Ia35f16956daa05e4e487bd4adf3d8db311788d48