Bug #34560694 	InnoDB: Some tests use too long socket argument for mysqld
If a path within the vardir is used to store a socket file, the resulting path for --socket param for mysqld executable may exceed 107 chars and lead to errors.
The --socket path must base on `$MYSQL_TMP_DIR` and have a filename not longer than `mysqld.1.sock` - a `tmp.sock` is used.

Change-Id: I20e29ee86718bc664ee28fd341d68bd0e353079e