WL#15307 MySQL Autopilot Auto Shape Predictor Server Component
The Shape Prediction functionality provides a shape suggestion
to improve the performance and latency of workload(s) running
on MySQL server. To do so, the functionality processes
MySQL-generated statistics while a MySQL DB System instance
is running a workload. Overall the functionality provides a buffer
pool size estimate and makes a recommendation for the most suitable
HeatWave Cloud Service and/or Oracle Cloud Infrastructure shape that
is available in customer's tenancy.

This workload estimate is based on the buffer pool usage and
performance in the MySQL server instance during the observation
period, and the growth or lack of growth in table size during the period.
The shape prediction function uses these statistics to estimate the
buffer pool size requirement for the workload.

Change-Id: If4a7d90ac0dd01cc421d2170ff7d6cc4462b5096