Bug #25584097: MYSQL_STMT_CLOSE HANGS WHEN THE STATEMENT IS CANCELED WITH KILL QUERY
Two problems:

Problem1:

The client library was not handling killing a cursor producing
mysql_stmt_execute being killed.
Basically it was not copying the error from net into the prepared
statement's error area and was not resetting the state when trying to read the result
and failing. Hence, at close, the client was trying to read the rest of
the resultset and failing.

Fixed by properly handling the error condition.

Problem2:
The server side materialized cursor wasn't closed properly when a KILL was sent
while the statement was still running and did not receive all of the data from the
cursor.

Fixed by making sure the cursor is properly disposed of.

Test case added.

Change-Id: Iff8305972948c4239bead8bd08b11088289a297e