Bug#34589153: Cleanup code in Item_func_like functions fix_fields() and resolve
Previously, class Item_func_like had a special fix_fields()
implementation. However, for the most part, this is not necessary.
The special handling can be moved to the resolve_type() function
instead. A specific implementation is still needed, however, but only
to set the can_use_prefix_key property before calling a generic
fix_fields() function.

Character set determination and propagation was carried out in two
inconsistent code blocks: Type propagation was based on the first
argument that had a string type, whereas aggregation of character set
was based on the logic in agg_arg_charsets_for_comparison().
In the new implementation, we first aggregated character set
based on all arguments (including the ESCAPE clause). Note that this
is compatible with existing implementation (except if ESCAPE causes
a difference). We then use the determined character set in type
propagation, instead of separately derive the character set to use here.

Notice the change of default collation for parameters to be the
connection collation. It was required to keep the same collation for
the second argument when the first argument was non-string.

Special handling of the ESCAPE clause is moved into resolve_type().

Change-Id: Ic09e62526cacde338fd74842a634b7eef69122db