BUG#34526001: Memory allocations going out of bound on               server upgrade
Description:
============
Upgrade from 5.7 to 8.0 with large number of tables in a
single database will consume huge amounts of memory.

Analysis:
=========
During the process of checking whether tables can be
upgraded, we fetch all the DD:Table objects upfront, iterate
over each one of them, do some checks and fetch their names
into a comma-separated list and perform CHECK TABLES ... FOR
UPGRADE on the entire list.

Fetching all objects upfront is not necessary in this case
and it contributes majorly to the memory consumption.

Fix:
====
We now fetch one object at a time, do some checks and fetch
its table name and release the object before proceeding with
the next object.

Change-Id: I4fbd919064b2b15707e45d0322be27e57b9fcb18