BUG#34357345: Do not use SQL Service API to set/get option super_read_only value
Group Replication needs to set/get some system options that are
inaccessible to plugins.
The 3 cases are:
 1) read_only
 2) super_read_only
 3) offline_mode
We also persist only the options:
 4) group_replication_single_primary_mode
 5) group_replication_enforce_update_everywhere_checks
To access those, Group Replication uses the SQL Service API which
requires to open a internal connection to the server and goes
through the complete SQL pipeline.
On 8.0 more efficient services were added to perform these
operations.

This patch does replace the use of the SQL Service API with the more
efficient services.

Change-Id: I0b3258ac1ff62e574f6a188936f64823b1a1d2c8