Bug#33294005: mysql server crash with fulltext index search [1/3, implicit grouping, noclose]
Some grouped queries that returned full-text search results could make
the server exit. This happened if the query execution plan used
streaming aggregation instead of temporary table aggregation and
failed to materialize the full-text search results before aggregation.

This patch fixes the bug for implicitly grouped queries. The optimizer
already had code to add the necessary materialization step before
aggregation, but it only did so for single-table queries. It is now
extended to also add it for multi-table queries, if needed.

Change-Id: I90018157e7faba669888ab7d74b344bd18e9dd43