Bug#34597056 MySQL Router crashes in the presence of execution of Nessus scan
When the connection is still in the x-protocol setup phase and client sends
valid but unexpected in this phase message (like SessionReset etc.) the Router
tries to forward it to the server even though the server connection is not yet
properly created and initialized. That leads to the null pointer dereference
and crash.

This patch fixes the issue by sending "Bad message" error back to the client
as a response to such message.

Change-Id: I48baf32b829f47d92fe1123a726f947d785bfc68