Bug #34565163: InnoDB: memcached.memc286_sdi_create fails due to content mismatch
Issue: Test memc286_sdi_create fails after patch for Bug #33673961: data
is incorrect after instant add column.

Cause: The current result file is incorrect, as it does not show that the table is in the system tablespace even when the table is not in file-per-table. The current output is the intended behaviour.

Fix: Re-recorded the test and updated the result file. Added another test table_options to avoid such errors in the future.

Change-Id: I39bd7967056697410c850a2dd1282530cabc7c71