Bug#32616816: WRONG RESULT WITH FULL-TEXT FUNCTION IN HAVING CLAUSE Bug#32934558: CRASH WITH MATCH() IN HAVING CONDITION
Evaluating the full-text search function (MATCH) after aggregation
does not work, because it needs the underlying scan to be positioned
on the correct row when it is invoked, and the position of underlying
scans is not preserved through aggregation.

If MATCH is called from the HAVING clause, it is evaluated after
aggregation, and the position of the underlying scan is lost. This
could lead to wrong results in some cases, and to the server exiting
in other cases.

This can be fixed by materializing the results from the calls to MATCH
before aggregation, so that the function does not have to be evaluated
again after aggregation. There is already code in the optimizer for
materializing the SELECT list before aggregation to make full-text
search work correctly. This patch makes the resolver extract the parts
of the HAVING clause that need materialization and add those
expressions as hidden elements in the SELECT list, so that they can be
materialized together with the SELECT list.

The hypergraph optimizer already had code to circumvent this issue,
and it was therefore not affected. Now that the issue is handled in
the resolver, most of the hypergraph optimizer's code for handling the
issue is no longer needed. The patch removes that code.

Change-Id: I32b6f9b430aeb8f0c3ddb936dfa9a824ac496df5