BUG#34607401: unhelpful 8.0.28->8.0.30 "Data Dictionary               initialization failed."
Description:
============
In 8.0.x, consider a stored procedure containing a SELECT
with optimizer hint 'MAX_EXECUTION_TIME' set to a value
higher than the allowed limit. Upgrading this data directory
to any higher 8.0 versions will cause the upgrade to abort.

Analysis:
=========
If there is a syntax error in an optimizer hint, the hint
will be ignored and a warning will be reported. This warning
will contain error code corresponding to ER_PARSE_ERROR. In
upgrade, when the statement with this optimizer hint is
parsed, the parser reports the warning and the error handler
considers this particular error code (ER_PARSE_ERROR) as an
error during parsing without considering the severity of the
SQL condition (Warning/Error/Note). Thus the upgrade is
aborted.

Fix:
====
The error handler now also checks the severity of the SQL
condition and considers only those conditions with SL_ERROR
as parse errors.

Change-Id: Id82d96cfadbde7b8a146db4e8d517f9f422b7805