Bug#34537146 client hangs if user sets the innodb-redo-log-capacity when in read-only mode
Description
 -----------
If user tries to set the innodb-redo-log-capacity when the server started
with --innodb-read-mode=ON then user connection hangs.

REDO log capacity modification needs involvement from log_files_governor_thread. But this log_files_governor_thread thread is a background thread. And background threads are not started if server is started in read_only mode. Thus this hang.

Fix:
----
Fix is not to allow changing the REDO capacity when server is started in read_only mode. Display error message to client and return if attempted.

Change-Id: I0827df97bb1d59a43f5393e3097fcbec50933c90