WL#15040 Add system variable to control EXPLAIN format.
Introduce a system variable 'explain_format' to specify a default
format to be used in EXPLAIN statement when the user does not specify
the format in the EXPLAIN statement. The compiled-in default value is
TRADITIONAL. This variable can be set either globally or in a session.

With hypergraph optimizer on, no format specified in EXPLAIN, and
explain_format set to TRADITIONAL, display the plan using TREE format.

With EXPLAIN ANALYZE, no format specified in EXPLAIN, and
explain_format set to TRADITIONAL, again use TREE format. This
preserves the existing behaviour.

Besides the usual format values, explain_format can have an additional
value "TRADITIONAL_STRICT", only for the sake of mtr tests. The mtr
tests would fail with hypergraph because of the silent conversion from
TRADITIONAL to TREE since that would result in plan outputs that would
differ in hypergraph. So in mtr now this variable explain_format will
be set to TRADITIONAL_STRICT, so that tests continue to show existing
behaviour of failing EXPLAIN statements with hypergraph and still
managing to pass because this particular error is ignored.

Change-Id: Ib6b1819ede619de94b3b551aae812caec418ba8e