Bug#34626727 Password for stored user is expired on remote
There are some complex situations involving RENAME USER that require
ndb_stored_grants.cc to pull apart the CREATE USER statement recorded
in a snapshot and rewrite it as a series of statements. There are
also some simple cases, when users are first created, where applying
the CREATE USER directly would succeed, and this would guarantee closer
equivalence between the originating mysql server and the other mysql
servers. The create_user() algorithm in ndb_stored_grants has never
tried to run the snapshot statement as-is to see if this is the case.

This patch changes that. Now ndb_stored_grants will attempt to run the
snapshot CREATE USER statement, and only fall back to the rewrite case
if that attempt fails.

Change-Id: Ia86269d5755597f0b0163f36a19327c35d5bb5a8