Bug#33574176 Compile MySQL with GCC 12
Compiler can not deduce that pointer can not be null pointer, adding an
explicit require that removeFirst succeed which in turn helps compiler
to understand that pointer is never null, which seems to be the trigger
of warning '... is out of bounds [0, 0] ...'.

Interesting is that the explict object destruction before memset, nor
the in place construction of object using same pointer compiler without
notice.

In function 'memset',
    inlined from 'TransientPool<Dbtc::GcpRecord, 8u>::release(Ptr<Dbtc::GcpRecord>)' at storage/ndb/src/kernel/vm/TransientPool.hpp:166:9,
    inlined from 'Dbtc::unlinkAndReleaseGcp(Ptr<Dbtc::GcpRecord>)' at storage/ndb/src/kernel/blocks/dbtc/DbtcMain.cpp:16694:26:
/usr/include/x86_64-linux-gnu/bits/string_fortified.h:59:33: error: 'memset' offset [0, 19] is out of the bounds [0, 0] [-Werror=array-bounds]
   59 |   return __builtin___memset_chk (__dest, __ch, __len,
      |          ~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~
   60 |                                  __glibc_objsize0 (__dest));
      |                                  ~~~~~~~~~~~~~~~~~~~~~~~~~~

Change-Id: I7461da32712ccde0043f9167379908749fc72b2b