Bug#33910786 Scalar subquery transformation combined with WHERE clause optimization lead to reject_multiple_rows being ineffective
Pushing down the condition into the derived table made from the
subquery leads to a wrong result here. If we rely run-time checking of
the cardinality, we should not push down the condition, since it will
potentially change the cardinality to zero or one, which is
semantically wrong:

This query

  SELECT 1 AS one FROM t WHERE 1=(SELECT 1 UNION SELECT 2)

is transformed to

SELECT 1 AS one FROM
       t JOIN
       ( SELECT 1 AS col1 UNION
         SELECT 2) derived
where 1 = derived.col1

and pushing down '1 = derived.col1' into the unio removes the
contribution from SELECT 2, which is wrong.

Solution: don't push down in this case.

Change-Id: If214f5fa88ed79c186ca4b9806445524e8f37e5e